Update WooCommerce 2.6.1

+++ web/app/plugins/woocommerce/includes/admin/class-wc-admin-taxonomies.php
@@ -38,13 +38,12 @@ class WC_Admin_Taxonomies {
 
 		// Taxonomy page descriptions
 		add_action( 'product_cat_pre_add_form', array( $this, 'product_cat_description' ) );
-		add_action( 'product_shipping_class_pre_add_form', array( $this, 'shipping_class_description' ) );
 
 		$attribute_taxonomies = wc_get_attribute_taxonomies();
 
-		if ( $attribute_taxonomies ) {
-			foreach ( array_keys( $attribute_taxonomies ) as $attribute ) {
-				add_action( $attribute . '_pre_add_form', array( $this, 'product_attribute_description' ) );
+		if ( ! empty( $attribute_taxonomies ) ) {
+			foreach ( $attribute_taxonomies as $attribute ) {
+				add_action( 'pa_' . $attribute->attribute_name . '_pre_add_form', array( $this, 'product_attribute_description' ) );
 			}
 		}
 
@@ -79,7 +78,7 @@ class WC_Admin_Taxonomies {
 
 		$term_id = absint( $term_id );
 
-		if ( $term_id ) {
+		if ( $term_id && get_option( 'db_version' ) < 34370 ) {
 			$wpdb->delete( $wpdb->woocommerce_termmeta, array( 'woocommerce_term_id' => $term_id ), array( '%d' ) );
 		}
 	}
@@ -89,7 +88,7 @@ class WC_Admin_Taxonomies {
 	 */
 	public function add_category_fields() {
 		?>
-		<div class="form-field">
+		<div class="form-field term-display-type-wrap">
 			<label for="display_type"><?php _e( 'Display type', 'woocommerce' ); ?></label>
 			<select id="display_type" name="display_type" class="postform">
 				<option value=""><?php _e( 'Default', 'woocommerce' ); ?></option>
@@ -98,7 +97,7 @@ class WC_Admin_Taxonomies {
 				<option value="both"><?php _e( 'Both', 'woocommerce' ); ?></option>
 			</select>
 		</div>
-		<div class="form-field">
+		<div class="form-field term-thumbnail-wrap">
 			<label><?php _e( 'Thumbnail', 'woocommerce' ); ?></label>
 			<div id="product_cat_thumbnail" style="float: left; margin-right: 10px;"><img src="<?php echo esc_url( wc_placeholder_img_src() ); ?>" width="60px" height="60px" /></div>
 			<div style="line-height: 60px;">
@@ -154,6 +153,24 @@ class WC_Admin_Taxonomies {
 					jQuery( '.remove_image_button' ).hide();
 					return false;
 				});
+
+				jQuery( document ).ajaxComplete( function( event, request, options ) {
+					if ( request && 4 === request.readyState && 200 === request.status
+						&& options.data && 0 <= options.data.indexOf( 'action=add-tag' ) ) {
+
+						var res = wpAjax.parseAjaxResponse( request.responseXML, 'ajax-response' );
+						if ( ! res || res.errors ) {
+							return;
+						}
+						// Clear Thumbnail fields on submit
+						jQuery( '#product_cat_thumbnail' ).find( 'img' ).attr( 'src', '<?php echo esc_js( wc_placeholder_img_src() ); ?>' );
+						jQuery( '#product_cat_thumbnail_id' ).val( '' );
+						jQuery( '.remove_image_button' ).hide();
+						// Clear Display type field on submit
+						jQuery( '#display_type' ).val( '' );
+						return;
+					}
+				} );
 
 			</script>
 			<div class="clear"></div>
@@ -279,13 +296,6 @@ class WC_Admin_Taxonomies {
 	/**
 	 * Description for shipping class page to aid users.
 	 */
-	public function shipping_class_description() {
-		echo wpautop( __( 'Shipping classes can be used to group products of similar type. These groups can then be used by certain shipping methods to provide different rates to different products.', 'woocommerce' ) );
-	}
-
-	/**
-	 * Description for shipping class page to aid users.
-	 */
 	public function product_attribute_description() {
 		echo wpautop( __( 'Attribute terms can be assigned to products and variations.<br/><br/><b>Note</b>: Deleting a term will remove it from all products and variations to which it has been assigned. Recreating a term will not automatically assign it back to products.', 'woocommerce' ) );
 	}
@@ -297,11 +307,14 @@ class WC_Admin_Taxonomies {
 	 * @return array
 	 */
 	public function product_cat_columns( $columns ) {
-		$new_columns          = array();
-		$new_columns['cb']    = $columns['cb'];
-		$new_columns['thumb'] = __( 'Image', 'woocommerce' );
+		$new_columns = array();
 
-		unset( $columns['cb'] );
+		if ( isset( $columns['cb'] ) ) {
+			$new_columns['cb'] = $columns['cb'];
+			unset( $columns['cb'] );
+		}
+
+		$new_columns['thumb'] = __( 'Image', 'woocommerce' );
 
 		return array_merge( $new_columns, $columns );
 	}
@@ -327,7 +340,7 @@ class WC_Admin_Taxonomies {
 			}
 
 			// Prevent esc_url from breaking spaces in urls for image embeds
-			// Ref: http://core.trac.wordpress.org/ticket/23605
+			// Ref: https://core.trac.wordpress.org/ticket/23605
 			$image = str_replace( ' ', '%20', $image );
 
 			$columns .= '<img src="' . esc_url( $image ) . '" alt="' . esc_attr__( 'Thumbnail', 'woocommerce' ) . '" class="wp-post-image" height="48" width="48" />';
